requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
cycler==0.10.0
Cython==0.29.23
fastapi==0.65.1
gensim==4.0.1
hdbscan==0.8.27
idna==2.10
joblib==1.0.1
kiwisolver==1.3.1
llvmlite==0.36.0
matplotlib==3.4.2
nltk==3.6.2
numba==0.53.1
numpy==1.20.3
pandas==1.2.4
Pillow==8.2.0
pydantic==1.8.2
pynndescent==0.5.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
smart-open==5.0.0
soupsieve==2.2.1
starlette==0.14.2
threadpoolctl==2.1.0
tqdm==4.60.0
typing-extensions==3.10.0.0
umap-learn==0.5.1
urllib3==1.26.4
